♻️ Trekk ut felles funksjoner for V1 og V2 beregningen for tilsyn barn
